Refactored risk assessment and risk assessment schedule editor screens so that they comply with model changes in the backend.
